# backend/openapi.yaml
openapi: 3.1.0
info:
  title: "TinyChat API"
  version: 0.1.0
  description: "API contract for the TinyChat backend."
security:
  - bearerAuth: []
paths:
...
  /api/delegations/status:
    get:
      operationId: getDelegationStatus
      summary: Check stored delegation status
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Delegation status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DelegationResponse"
              examples:
                active:
                  value:
                    status: active
                    expiresAt: "2026-05-20T12:00:00.000Z"
                expired:
                  value:
                    status: expired
                    expiresAt: "2026-05-19T12:00:00.000Z"
                stale:
                  value:
                    status: stale
                    expiresAt: null
                none:
                  value:
                    status: none
                    expiresAt: null
        "401":
          $ref: "#/components/responses/Unauthenticated"
  /api/delegations:
    post:
      operationId: storeDelegation
      summary: Store and activate a portable delegation
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [serialized]
              properties:
                serialized:
                  type: string
      responses:
        "200":
          description: Delegation active
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DelegationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
    delete:
      operationId: revokeDelegation
      summary: Remove stored delegation
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Delegation removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DelegationResponse"
        "401":
          $ref: "#/components/responses/Unauthenticated"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Backend session token returned by SIWE verification.
  schemas:
    ApiError:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
...
    DelegationStatus:
      type: string
      enum: [active, expired, none, stale]
    DelegationResponse:
      type: object
      required: [status, expiresAt]
      properties:
        status:
          $ref: "#/components/schemas/DelegationStatus"
        expiresAt:
          type:
            - string
            - "null"
          format: date-time
        verified:
          type: boolean
          description: >-
            Whether the stored delegation's `att` has been cryptographically verified. Present
            and `false` on every `active` response today: the §9.2 verification chokepoint does
            not exist yet, so `active` means stored, unexpired and addressed to this backend —
            the resources, delegator and expiry behind it all come from the bundle's unsigned
            top-level fields. Additive on purpose; the `status` enum is unchanged.
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    Unauthenticated:
      description: Authentication failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
...